implemented link management

+++ slepeweb-cms/src/main/java/com/slepeweb/cms/service/ItemServiceImpl.java
@@ -190,9 +190,9 @@ public class ItemServiceImpl extends BaseServiceImpl implements ItemService {
 	private void removeStaleLinks(List<Link> dbRecordLinks, List<Link> updatedLinks) {
 		if (dbRecordLinks != null && updatedLinks != null) {
 			for (Link dbLink : dbRecordLinks) {
-				if (! updatedLinks.contains(dbLink)) {
+				if (! updatedLinks.contains(dbLink) && dbLink.getType() != LinkType.binding) {
 					dbLink.delete();
-					LOG.info(compose("Deleted old child link", dbLink));
+					LOG.info(compose("Deleted old inline/relation link", dbLink));
 				}
 			}
 		}
